pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "garmin-analytics"
version = "0.1.0"
description = "Garmin export analytics toolkit"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]
dependencies = [
  "pandas",
  "pyarrow",
  "orjson",
  "rich",
  "typer",
  "python-dotenv",
]

[project.optional-dependencies]
test = ["pytest"]

[project.scripts]
garmin-analytics = "garmin_analytics.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
